pyproject.toml
[tool.poetry]
name = "workout-discord-bot"
version = "0.1.0"
description = "Discord bot for workout tracking and penalty calculation"
authors = ["arsture"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
discord-py = "^2.5.2"
python-dotenv = "^1.1.0"
apscheduler = "^3.11.0"
supabase = "^2.15.3"
pytz = "^2025.2"
flask = "^3.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
pytest-xdist = "^3.6.0"
black = "^24.0.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
